backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==3.0.1
gunicorn==21.2.0
marshmallow==3.20.1
python-dotenv==1.0.0
Pillow==10.1.0
python-magic==0.4.27